requirements.txt
asgiref==3.8.1
Django==5.1.6
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
